refactor: refactor the code to use zustand and move some interface to @type  and rename hooks to .ts from .tsx
