+++ src/features/chat/components/chat-input-box.tsx
@@ -5,8 +5,8 @@ import { Textarea } from "@/components/ui/textarea"
 import { useAutoResizeTextarea } from "@/hooks/use-auto-resize-textarea"
 import { cn } from "@/lib/utils"
 import SendOrStopButton from "@/features/chat/components/send-or-stop-button"
-import { useChatInput } from "@/features/chat/hooks/use-chat-input"
-import { useLoadStream } from "@/features/chat/hooks/use-load-stream"
+import { useChatInput } from "@/features/chat/stores/chat-input-store"
+import { useLoadStream } from "@/features/chat/stores/load-stream-store"
 import ModelMenu from "@/features/model/components/model-menu"
 import PromptSelectorDialog from "@/features/prompt/components/prompt-selector-dialog"
 import TabsSelect from "@/features/tabs/components/tabs-select"
