+++ src/features/chat/components/chat-message-content.tsx
@@ -2,7 +2,7 @@ import { MarkdownRenderer } from "@/components/markdown-renderer"
 import { cn } from "@/lib/utils"
 import ChatMessageLoadingIndicator from "@/features/chat/components/chat-message-loading-indicator"
 import ChatMessageMetricsDropdown from "@/features/chat/components/chat-message-metrics-dropdown"
-import { useLoadStream } from "@/features/chat/hooks/use-load-stream"
+import { useLoadStream } from "@/features/chat/stores/load-stream-store"
 import type { ChatMessage } from "@/types"
 
 export default function ChatMessageContent({
