+++ src/features/chat/stores/chat-input-store.ts
@@ -1,11 +1,18 @@
 import { create } from "zustand"
+import { useShallow } from "zustand/react/shallow"
 
-interface ChatInput {
-  input: string
-  setInput: (value: string) => void
-}
+import type { ChatInput } from "@/types"
 
 export const chatInputStore = create<ChatInput>((set) => ({
   input: "",
   setInput: (text) => set({ input: text })
 }))
+
+export const useChatInput = () => {
+  return chatInputStore(
+    useShallow((s) => ({
+      input: s.input,
+      setInput: s.setInput
+    }))
+  )
+}
