+++ src/features/model/hooks/use-model-info.ts
