+++ src/features/sessions/components/chat-session-selector.tsx
@@ -7,7 +7,7 @@ import { Button } from "@/components/ui/button"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet"
 import { cn } from "@/lib/utils"
-import { useChatSessions } from "@/features/sessions/hooks/use-chat-sessions"
+import { useChatSessions } from "@/features/sessions/stores/chat-session-store"
 
 export default function ChatSessionSelector() {
   const {
